Refactor hermes-logger crate
