+++ crates/cosmos/cosmos-integration-tests/src/contexts/binary_channel/test_driver.rs
@@ -5,9 +5,9 @@ use cgp::core::field::Index;
 use cgp::prelude::*;
 use hermes_error::handlers::debug::DebugError;
 use hermes_error::impls::ProvideHermesError;
-use hermes_logger::ProvideHermesLogger;
+use hermes_logger::UseHermesLogger;
 use hermes_logging_components::traits::has_logger::{
-    GlobalLoggerGetterComponent, LoggerGetterComponent, LoggerTypeComponent,
+    GlobalLoggerGetterComponent, LoggerGetterComponent, LoggerTypeProviderComponent,
 };
 use hermes_relayer_components::multi::traits::birelay_at::BiRelayTypeAtComponent;
 use hermes_relayer_components::multi::traits::chain_at::ChainTypeAtComponent;
@@ -55,11 +55,11 @@ delegate_components! {
         ]:
             ProvideCosmosTestTypes,
         [
-            LoggerTypeComponent,
+            LoggerTypeProviderComponent,
             LoggerGetterComponent,
             GlobalLoggerGetterComponent,
         ]:
-            ProvideHermesLogger,
+            UseHermesLogger,
     }
 }
