+++ crates/logging/logging-components/src/traits/has_logger.rs
@@ -1,22 +1,21 @@
 use cgp::prelude::*;
 
-#[cgp_component {
-  name: LoggerTypeComponent,
-  provider: ProvideLoggerType,
+#[cgp_type {
+    provider: LoggerTypeProvider,
 }]
 pub trait HasLoggerType: Async {
     type Logger: Async;
 }
 
 #[cgp_component {
-  provider: LoggerGetter,
+    provider: LoggerGetter,
 }]
 pub trait HasLogger: HasLoggerType {
     fn logger(&self) -> &Self::Logger;
 }
 
 #[cgp_component {
-  provider: GlobalLoggerGetter,
+    provider: GlobalLoggerGetter,
 }]
 pub trait HasGlobalLogger: HasLoggerType {
     fn global_logger() -> &'static Self::Logger;
