+++ crates/mock/mock-relayer/src/relayer_mock/components/chain.rs
@@ -1,7 +1,7 @@
 use cgp::prelude::*;
 use hermes_logging_components::contexts::no_logger::ProvideNoLogger;
 use hermes_logging_components::traits::has_logger::{
-    GlobalLoggerGetterComponent, LoggerGetterComponent, LoggerTypeComponent,
+    GlobalLoggerGetterComponent, LoggerGetterComponent, LoggerTypeProviderComponent,
 };
 use hermes_relayer_components::chain::impls::types::ack::ProvideBytesAcknowlegement;
 use hermes_relayer_components::chain::traits::types::packets::ack::AcknowledgementTypeComponent;
@@ -11,7 +11,7 @@ pub use crate::relayer_mock::contexts::chain::MockChainComponents;
 delegate_components! {
     MockChainComponents {
         [
-            LoggerTypeComponent,
+            LoggerTypeProviderComponent,
             LoggerGetterComponent,
             GlobalLoggerGetterComponent,
         ]:
